Move files to components/login
